refactor: move helper functions to Parser class
